Moves Collision event handlers to physics_system
