+++ src/pyrite/physics/physics_service.py
@@ -1,20 +1,18 @@
 from __future__ import annotations
 
-from typing import Any, TypeAlias, TYPE_CHECKING
+from typing import TypeAlias, TYPE_CHECKING
 
 import cffi
 
 import pymunk
 
-from .. import physics  # COMPONENT_TYPE, get_collider_components
+# from .. import physics  # COMPONENT_TYPE, get_collider_components
 
 if TYPE_CHECKING:
     from pymunk import (
-        Arbiter,
         PointQueryInfo,
         SegmentQueryInfo,
         ShapeFilter,
-        Space,
     )
 
 
@@ -27,34 +25,9 @@ Point: TypeAlias = tuple[float, float]
 COMPONENT_TYPE: int = 2 ** (cffi.FFI().sizeof("int") * 8 - 1) - 1
 
 
-# Figure out where to put this so it doesn't cause circular imports
-# Track _bodies here? Would remove RigidbodyComponent import
-def post_solve(arbiter: Arbiter, space: Space, data: Any):
-    collider1, collider2 = physics.get_collider_components(arbiter)
-    if arbiter.is_first_contact:
-        if collider1.compare_mask(collider2):
-            collider1.OnTouch(collider1, collider2)
-        if collider2.compare_mask(collider1):
-            collider2.OnTouch(collider2, collider1)
-    if collider1.compare_mask(collider2):
-        collider1.WhileTouching(collider1, collider2)
-    if collider2.compare_mask(collider1):
-        collider2.WhileTouching(collider2, collider1)
-
-
-def separate(arbiter: Arbiter, space: Space, data: Any):
-    collider1, collider2 = physics.get_collider_components(arbiter)
-    if collider1.compare_mask(collider2):
-        collider1.OnSeparate(collider1, collider2)
-    if collider2.compare_mask(collider1):
-        collider2.OnSeparate(collider2, collider1)
-
-
 class PhysicsService:
     space = pymunk.Space()
     comp_handler = space.add_collision_handler(COMPONENT_TYPE, COMPONENT_TYPE)
-    comp_handler.post_solve = post_solve
-    comp_handler.separate = separate
 
     @staticmethod
     def cast_ray(
